requirements.txt
flask
numpy
python-dotenv
langchain-community
langchain-openai
chromadb
openai
optional
